API-25989 Add watermarked downloads to appeals APIs
Adds a PDF download route to the Higher-Level Reviews, Supplemental
Claims, and Notice of Disagreements APIs at the endpoint
`/services/appeals/<appeal-type>/v0/forms/<number>/:id/download`.
This allows the user to download a watermarked version of the PDF for
an appeal that is submitted, provided that the PII has not yet been deleted
from the server.