Implement service for returning missing IDs for User
Implements a basic service class that checks for the presence of certain key User identifiers required by the 526 Form flow and returns any that are missing as a list. This behavior will be required by improved error messaging in the front end which prompts a user missing any of these ids to call a contact center and provide them with the list of what is missing. Currently this messaging is based on Pundit-based 'service' authorization that lives in app/services/users/services.rb and app/policies/evss_policy.rb. Because the front end currently only gets a list of abstracted 'services' a User has access to based on this data (e.g. 'evss_claims') instead of which specific identifiers may be missing, we can't currently provide the user with that information to pass on to a contact center.

This service will be used in a new endpoint that will be queried when a User begins the 526 Form Flow.

Note many of these identifiers aren't attributes on a User record, they're delegated to other objects, namely MPIData and a User's 'profile'