Refactor service to return hash mapping of identifiers
Refactors the service method in Users::Form526UserIdentificationVerifier to return a mapping of required 526 identifiers to a boolean noting if they are present or not.

We may not want to serialize all of the actual values for these identifiers as they may contain sensitive personal information. But we want to be aware of their presence or lack thereof for Form 526. This service will be used when serializing this information in the user profile metadata produced by app/services/users/profile.rb.

We are already validating the presence of all of these identifiers in app/policies/evss_policy.rb, but we need to surface which ones may be missing so the user can know that information.