Include Form 526 Identifiers presence in Pundit policy, profile
Includes the resulting hash from the Form526UserIdentificationVerifier in the claims section of the User profile that is returned by the /v0/user endpoint.

This is referenced via a new Pundit policy, Form526Policy, so this behavior is in the same place as other access checks we store in these policies. However, unlike other policies with a binary access? method which returns true or false, this policy needs to return the list of identifiers and whether we have them or not. This is so the front end can display the missing identifiers and the user can reference them when calling a Contact Center.